+++ source/main/include/cbase/c_binary_search.h
@@ -7,14 +7,14 @@
 
 namespace ncore
 {
-	typedef	s32		(*compare_predicate_fn)(const void* inItem, const void* inData, s32 inIndex);
+	typedef	s8		(*compare_predicate_fn)(const void* inItem, const void* inData, s64 inIndex);
 
 	// Exact search, behaves like g_LowerBound by returning the index to the lowest equal data (lower bound) and not
 	// just the first equal data that is encountered.
-	extern u32		g_BinarySearch(const void* inItem, const void* inData, u32 inLength, compare_predicate_fn predicate);
+	extern s64		g_BinarySearch(const void* inItem, const void* inData, u64 inLength, compare_predicate_fn predicate);
 
-	extern u32		g_LowerBound(const void* inItem, const void* inData, u32 inLength, compare_predicate_fn predicate);
-	extern u32		g_UpperBound(const void* inItem, const void* inData, u32 inLength, compare_predicate_fn predicate);
+	extern s64		g_LowerBound(const void* inItem, const void* inData, u64 inLength, compare_predicate_fn predicate);
+	extern s64		g_UpperBound(const void* inItem, const void* inData, u64 inLength, compare_predicate_fn predicate);
 };
 
 #endif	///< __CCORE_BINARY_SEARCH_H__
